refactor(users-service): replace methods related to friendship between users
