+++ src/services/users-service.ts
@@ -1,8 +1,8 @@
 import { prisma } from '../db';
-import { BadRequestError, NotFoundError } from '../errors/app-errors';
+import { NotFoundError } from '../errors/app-errors';
 import { prismaErrorHandler } from '../errors/prisma-error-handler';
 import { Nullable } from '../types/global';
-import { UsersCategories } from '../types/responses';
+import { UsersCategoriesCount } from '../types/responses';
 
 class UsersService {
   private userSelect: { id: boolean; name: boolean; email: boolean; image: boolean };
@@ -46,10 +46,7 @@ class UsersService {
                 },
                 take: 1,
               },
-              users: {
-                where: { id: { not: id } },
-                select: this.userSelect,
-              },
+              users: { select: this.userSelect },
               _count: {
                 select: {
                   messages: { where: { isRead: false } },
@@ -89,7 +86,7 @@ class UsersService {
     }
   }
 
-  async getCategoriesCounts(id: string) {
+  async getCategoriesCount(id: string) {
     try {
       const allUsersCountPromise = prisma.user.count();
       const restCategoriesCountsPromise = prisma.user.findUnique({
@@ -110,12 +107,14 @@ class UsersService {
         restCategoriesCountsPromise,
       ]);
 
-      const counts: { [Property in UsersCategories]: Nullable<number> } = {
-        all: allUsersCount,
-        friends: restCategoriesCount?._count.friends ?? null,
-        incomingRequests: restCategoriesCount?._count.incomingRequests ?? null,
-        outcomingRequests: restCategoriesCount?._count.outcomingRequests ?? null,
-      };
+      const counts: Nullable<UsersCategoriesCount> = restCategoriesCount
+        ? {
+            all: allUsersCount,
+            friends: restCategoriesCount._count.friends,
+            incomingRequests: restCategoriesCount._count.incomingRequests,
+            outcomingRequests: restCategoriesCount._count.outcomingRequests,
+          }
+        : null;
 
       return counts;
     } catch (err) {
@@ -169,160 +168,6 @@ class UsersService {
       } else {
         throw new NotFoundError('user', { id });
       }
-    } catch (err) {
-      prismaErrorHandler(err);
-    }
-  }
-
-  async sendFriendRequest({ senderId, receiverId }: { senderId: string; receiverId: string }) {
-    try {
-      const isRequestAlreadySent = Boolean(
-        await prisma.user.findUnique({
-          where: {
-            id: senderId,
-            OR: [
-              { outcomingRequests: { some: { id: receiverId } } },
-              { incomingRequests: { some: { id: receiverId } } },
-              { friends: { some: { id: receiverId } } },
-            ],
-          },
-        })
-      );
-
-      if (isRequestAlreadySent) {
-        throw new BadRequestError(
-          'Friend request between the sender and the receiver exists (or has been accepted already)'
-        );
-      }
-
-      const updatedSender = await prisma.user.update({
-        where: { id: senderId },
-        data: { outcomingRequests: { connect: { id: receiverId } } },
-        select: {
-          ...this.userSelect,
-          _count: {
-            select: {
-              friends: { where: { id: senderId } },
-              incomingRequests: { where: { id: senderId } },
-              outcomingRequests: { where: { id: senderId } },
-            },
-          },
-        },
-      });
-
-      return updatedSender;
-    } catch (err) {
-      prismaErrorHandler(err);
-    }
-  }
-
-  async acceptFriendRequest({
-    requestSenderId,
-    requestReceiverId,
-  }: {
-    requestSenderId: string;
-    requestReceiverId: string;
-  }) {
-    try {
-      const updateRequestSenderPromise = prisma.user.update({
-        where: { id: requestSenderId },
-        select: this.userSelect,
-        data: {
-          outcomingRequests: { disconnect: { id: requestReceiverId } },
-          friends: { connect: { id: requestReceiverId } },
-        },
-      });
-
-      const updateRequestReceiverPromise = prisma.user.update({
-        where: { id: requestReceiverId },
-        select: this.userSelect,
-        data: {
-          incomingRequests: { disconnect: { id: requestSenderId } },
-          friends: { connect: { id: requestSenderId } },
-        },
-      });
-
-      const createOrEnableCommonChatPromise = prisma.chat
-        .updateMany({
-          where: { users: { every: { id: { in: [requestReceiverId, requestSenderId] } } } },
-          data: { isDisabled: false },
-        })
-        .then(({ count: commonChatCount }) => {
-          if (!commonChatCount) {
-            prisma.chat.create({
-              data: {
-                users: { connect: [{ id: requestSenderId }, { id: requestReceiverId }] },
-              },
-            });
-          }
-        });
-
-      const [_, updatedRequestReceiver] = await Promise.all([
-        updateRequestSenderPromise,
-        updateRequestReceiverPromise,
-        createOrEnableCommonChatPromise,
-      ]);
-
-      return updatedRequestReceiver;
-    } catch (err) {
-      prismaErrorHandler(err);
-    }
-  }
-
-  async refuseFriendRequest({
-    requestSenderId,
-    requestReceiverId,
-  }: {
-    requestSenderId: string;
-    requestReceiverId: string;
-  }) {
-    try {
-      const updatedRequestReceiver = await prisma.user.update({
-        where: { id: requestReceiverId },
-        select: this.userSelect,
-        data: { incomingRequests: { disconnect: { id: requestSenderId } } },
-      });
-
-      return updatedRequestReceiver;
-    } catch (err) {
-      prismaErrorHandler(err);
-    }
-  }
-
-  async removeFromFriends({ userId, friendId }: { userId: string; friendId: string }) {
-    try {
-      const hasUserTheFriend = Boolean(
-        await prisma.user.findUnique({
-          where: {
-            id: userId,
-            friends: { some: { id: friendId } },
-          },
-        })
-      );
-
-      if (!hasUserTheFriend) {
-        throw new NotFoundError('user', { id: userId, friends: [{ id: friendId }] });
-      }
-
-      const updateUserPromise = prisma.user.update({
-        where: { id: userId },
-        select: this.userSelect,
-        data: {
-          friends: { disconnect: { id: friendId } },
-          friendOf: { disconnect: { id: friendId } },
-        },
-      });
-
-      const disableCommonChatPromise = prisma.chat.updateMany({
-        where: {
-          AND: [{ users: { some: { id: userId } } }, { users: { some: { id: friendId } } }],
-        },
-        data: { isDisabled: true },
-      });
-
-      const [updatedUser] = await Promise.all([updateUserPromise, disableCommonChatPromise]);
-
-      return updatedUser;
     } catch (err) {
       prismaErrorHandler(err);
     }
